pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qrc-ev"
version = "0.1.0"
description = "Quantum Reservoir Computing for Electric Vehicle Charging Demand Forecasting"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "QRC-EV Research Team"}
]
keywords = ["quantum computing", "reservoir computing", "ev charging", "forecasting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pennylane>=0.39",
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "hypothesis>=6.0",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "mypy>=1.5",
    "black>=23.0",
    "isort>=5.12",
]
cudaq = [
    "cuda-quantum>=0.9",
]
gpu = [
    "cuda-quantum>=0.9",
    "pennylane-lightning-gpu>=0.39",
]

[project.urls]
Homepage = "https://github.com/qrc-ev-research/qrc-ev"
Documentation = "https://github.com/qrc-ev-research/qrc-ev/docs"
Repository = "https://github.com/qrc-ev-research/qrc-ev"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src/qrc_ev",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "property: property-based tests (hypothesis)",
    "integration: end-to-end integration tests",
    "slow: tests that take >5 seconds",
    "gpu: tests that require GPU (CUDA-Quantum)",
]

[tool.coverage.run]
source = ["src/qrc_ev"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
